MultiplePDFs/requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2
faiss-cpu # vector database
langchain_google_genai 